+++ src/main/java/com/example/workshop/springboot/repository/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 // Norberto Enomoto
-public interface UserRepository extends CrudRepository<UserModel, Integer>  {  
+public interface UserRepository extends CrudRepository<UserModel, Long>  {  
     Iterable<UserModel> findByName(String name);
 }
